% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_vertical}
\alias{build_vertical}
\title{Build vertical project}
\usage{
build_vertical(clean = TRUE, ...)
}
\arguments{
\item{clean}{logical, when clean=TRUE (the default), the \code{docs} folder is cleaned (e.g., completely wiped) using \code{pkgdown::clean_site()}, otherwise when clean=FALSE \code{clean_site()} will not be run.}

\item{...}{params, pass additional parameters to \code{pkgdown::build_site(...)}}
}
\description{
Build the website associated with a vertical project. This is mostly a wrapper to \code{pkgdown::build_site()}, but extended to render .Rmd content from other folders.
}
\section{Usage}{
\preformatted{build_vertical()
}
}

