% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{document_data}
\alias{document_data}
\title{Add R dataframe and document}
\usage{
document_data(...)
}
\arguments{
\item{...}{the name of a single dataframe in the global environment}
}
\value{
files, adds dataframe as .rda to data folder, adds .R to R for documentation
}
\description{
A wrapper to usethis::use_data(), usethis::use_r(), and sinew::makeOxygen()
}
\details{
document_data() is a helper function for quickly adding an existing dataframe to an R package. For example, if mydf was an existing data frame, document_data(mydf) would create the data folder (if it doesn't exist), add mydf.rda to the data folder, add mydf.R to the R folder with a roxygen skeleton for documenting the data, and open mydf.R for editing.
}
\examples{
\dontrun{
document_data(mydf)
}
}
