% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_jspsych}
\alias{init_jspsych}
\title{Initialize jsPsych experiment}
\usage{
init_jspsych()
}
\description{
Initialize a jsPsych experiment in the \code{experiments} folder of a vertical project
}
\details{
This function does the following:
\enumerate{
\item creates the \code{experiments} folder
\item Downloads the most recent `jspsych`` library from \url{https://github.com/jspsych/jsPsych/releases}
\item Adds a minimal \verb{jspsych`` example experiment to }experiments/experiment-1/`
}

See the \href{https://www.jspsych.org}{jspsych documentation} for more information about using jspsych to build behavioral experiments for the web.
}
