% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_jspsychr}
\alias{init_jspsychr}
\title{Initialize jspsychr experiment}
\usage{
init_jspsychr()
}
\description{
Initialize a jspsychr template in the \code{experiments} folder of a vertical project
}
\details{
This function does the following:
\enumerate{
\item creates the \code{experiments} folder
\item Adds a \code{jspsychr} template (Experiment_1), which is an example of using R Studio and R Markdown to author a \code{jspsych} experiment
}

See the \href{https://www.jspsych.org}{jspsych documentation} for more information about using jspsych to build behavioral experiments for the web.

See the \href{https://crumplab.github.io/jspsychr/}{jspsychr documentation} for more information about using R Markdown to write \code{jspsych} experiments.
}
