% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_papaja}
\alias{init_papaja}
\title{Initialize manuscript}
\usage{
init_papaja()
}
\description{
Initialize \code{papaja} R Markdown APA manuscript in the \code{manuscript} folder.
}
\details{
Run this function to add a \code{papaja} manuscript component to a vertical project at a later time (assuming it wasn't created by \code{vertical_project()} or \code{init_vertical_project()} during initialization.)

See the \href{https://crsh.github.io/papaja_man/}{papaja documentation} for more information.
}
