% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_poster}
\alias{init_poster}
\title{Initialize posters}
\usage{
init_poster()
}
\description{
Initialize R Markdown \code{posterdown} template in \code{posters} folder of a vertical project
}
\details{
Run on initialization, and can be run from the console to include a posters folder and poster template at a later time. See the \href{https://github.com/brentthorne/posterdown}{posterdown documentation} for more information. Note that \code{vertical} loads one of the three possible \code{posterdown} templates.
}
