% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_slides}
\alias{init_slides}
\title{Initialize slides}
\usage{
init_slides()
}
\description{
Initialize slidy R Markdown template in slides folder of a vertical project
}
\details{
Run on initialization, and can be run from the console to include a slides folder and slidy template at a later time. See the \href{https://bookdown.org/yihui/rmarkdown/slidy-presentation.html}{slidy documentation} for additional information. There are other R markdown slide templates not suggested by vertical, but that are very good alternatives (e.g., \code{xaringan}). Simply add your template of choice to the slides folder.
}
