% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_supplemental}
\alias{init_supplemental}
\title{Initialize supplemental materials}
\usage{
init_supplemental(prjct_name = NULL)
}
\description{
Initialize Rmarkdown SOM in the appropriate location of a vertical project
}
