% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_vertical_project}
\alias{init_vertical_project}
\title{Initialize a vertical project inside an existing R project}
\usage{
init_vertical_project(
  init_git = TRUE,
  init_ms = TRUE,
  init_som = TRUE,
  init_slides = TRUE,
  init_poster = TRUE,
  init_exp = TRUE
)
}
\arguments{
\item{init_git}{logical, enables git, init_git=TRUE by default}

\item{init_ms}{logical, enables manuscript with papaja, init_ms=TRUE by default}

\item{init_som}{logical, enables supplementary with vignettes, init_som=TRUE by default}

\item{init_slides}{logical, enables slides with slidy, init_slides=TRUE by default}

\item{init_poster}{logical, enables posters with posterdown, init_poster=TRUE by default}

\item{init_exp}{logical, enables experiments with jspsych, init_exp=TRUE by default}
}
\value{
files, a vertical project template inside an existing R project
}
\description{
If you are working inside an existing R project, typically an empty one, or perhaps a project in RStudio cloud, use this function to initialize a vertical project inside the existing project. The name of the existing project (e.g., the parent folder name) must be a valid R package name (numbers, letters, and periods, but no periods at the end, and no spaces, dashes, or underscores).
}
