% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_folder_to_docs.R
\name{render_folder_to_docs}
\alias{render_folder_to_docs}
\title{Render rmd ina folder to docs}
\usage{
render_folder_to_docs(folder_name, dest_folder = "docs")
}
\arguments{
\item{folder_name}{string, name of folder , eg "slides"}

\item{dest_folder}{string, name of destination folder, default = "docs}
}
\value{
rendered rmd files in destination folder
}
\description{
Render rmd ina folder to docs
}
