% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain.R
\name{suggest_yml}
\alias{suggest_yml}
\title{Suggest yml for _pkgdown.yml}
\usage{
suggest_yml()
}
\value{
yml copied to console
}
\description{
Run \code{vertical::suggest_yml} to generate a suggestions for populating \verb{_pkgdown.yml}. The suggested yml is written to the console. Copy this to \verb{_pkgdown.yml} and modify as necessary.
}
\details{
\verb{_pkgdown.yml} can be further modified by hand to achieve various customizations to the website. See the \href{https://pkgdown.r-lib.org/reference/build_site.html}{pkgdown documentation} for additional information.
}
