% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain.R
\name{update_yml}
\alias{update_yml}
\title{Update _pkgdown.yml for Rmds in vertical folders}
\usage{
update_yml(vertical_folder, docs_folder, yml_component, tab_title)
}
\arguments{
\item{vertical_folder}{character, name of a vertical folder, e.g., "vignettes"}

\item{docs_folder}{character, name of folder in docs, e.g., "articles"}

\item{yml_component}{character, name of the corresponding yml component, e.g., "articles"}

\item{tab_title}{character, the title of the component tab written to the website, e.g., "Supplementary"}
}
\value{
file, an updated _pkgdown.yml file
}
\description{
Write yml for the website navigation bar; run automatically by \code{build_vertical()}.
}
\details{
\code{update_yml()} is a helper function for adding vertical content to the website navigation bar by upating \verb{_pkgdown.yml}.

Specifically, the titles of .Rmds in  \code{vertical_folder} and its subfolders are listed under the associated navbar component in \verb{_pkgdown.yml}. If subfolders exist, then the name of the subfolder is used as a section header in the list. The sections and titles are shown in the dropdown navigation tab on the website.

\verb{_pkgdown.yml} can be further modified by hand to achieve various customizations to the website. See the \href{https://pkgdown.r-lib.org/reference/build_site.html}{pkgdown documentation} for additional information.
}
