% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_vertical_project}
\alias{init_vertical_project}
\title{Initialize vertical project from command line}
\usage{
init_vertical_project(
  project_name = NULL,
  project_path = NULL,
  init_git = NULL,
  init_ms = NULL,
  init_som = NULL,
  init_slides = NULL,
  init_poster = NULL,
  init_exp = NULL
)
}
\arguments{
\item{project_name}{character, name of new vertical project, default is project_name = NULL, for initializing inside an existing empty R studio project}

\item{project_path}{character, path where new project should be created, default is project_path = NULL for initializing inside an existing empty R studio project}

\item{init_git}{logical, enables git, init_git=TRUE by default}

\item{init_ms}{logical, enables manuscript with papaja, init_ms=TRUE by default}

\item{init_som}{logical, enables supplementary with vignettes, init_som=TRUE by default}

\item{init_slides}{logical, enables slides with slidy, init_slides=TRUE by default}

\item{init_poster}{logical, enables posters with posterdown, init_poster=TRUE by default}

\item{init_exp}{logical, enables experiments with jspsych, init_exp=TRUE by default}
}
\value{
\strong{files}, a vertical project template inside an existing or new R Studio project
}
\description{
Use \code{vertical::init_vertical_project()} to initialize a vertical project from the command line.
}
\details{
\strong{Interactive: Get asked what you want to do}\preformatted{vertical::init_vertical_project()
}

This opens a dialogue in the console. You choose whether to make a brand new RStudio project, or to initialize vertical in an existing project that you are already working. Follow the prompts and choose which vertical modules to load.

\strong{Scripted: Do what you want to do}

If any input parameters are specified, then initialization is done directly without a dialogue.

\strong{Initializing a vertical project inside an existing project}:
It is not necessary to name the project, because a named project should already be open. At least one of the modules should be specified. Default NULL values are set to FALSE, and those modules are not loaded.\preformatted{init_vertical_project(init_git=TRUE,...)
}

The name of the existing project (e.g., the project folder name) must be a valid R package name (numbers, letters, and periods, but no periods at the end, and no spaces, dashes, or underscores). It is not necessary to supply project_name or project_path, they default to the current existing project.

\strong{Initializing a vertical project as a new project}:
Provide a name and path to set up a new vertical project as an R studio project, and specify which modules should be loaded by setting their parameters to TRUE.\preformatted{init_vertical_project(project_name = "yourname",
                      project_path = "~/Desktop/",
                      init_git = TRUE,
                      ...)
}

For example, the above creates the folder \verb{~Desktop/yourname/}, creates the vertical project template in that folder, and opens a new RStudio session with the new project loaded. Make sure \code{yourname} is valid R package name.
}
