% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{init_vertical_project}
\alias{init_vertical_project}
\title{Initialize a vertical project from command line}
\usage{
init_vertical_project(
  project_name = NULL,
  project_path = NULL,
  init_git = TRUE,
  init_ms = TRUE,
  init_som = TRUE,
  init_slides = TRUE,
  init_poster = TRUE,
  init_exp = TRUE
)
}
\arguments{
\item{project_name}{character, name of new vertical project, default is project_name = NULL, for initializing inside an existing empty R studio project}

\item{project_path}{character, path where new project should be created, default is project_path = NULL for initializing inside an existing empty R studio project}

\item{init_git}{logical, enables git, init_git=TRUE by default}

\item{init_ms}{logical, enables manuscript with papaja, init_ms=TRUE by default}

\item{init_som}{logical, enables supplementary with vignettes, init_som=TRUE by default}

\item{init_slides}{logical, enables slides with slidy, init_slides=TRUE by default}

\item{init_poster}{logical, enables posters with posterdown, init_poster=TRUE by default}

\item{init_exp}{logical, enables experiments with jspsych, init_exp=TRUE by default}
}
\value{
files, a vertical project template inside an existing or new R Studio project
}
\description{
Initiliazing a vertical project inside an existing project: The name of the existing project (e.g., the project folder name) must be a valid R package name (numbers, letters, and periods, but no periods at the end, and no spaces, dashes, or underscores). It is not necessary to supply project_name or project_path, they default to the current existing project.

Initializing a vertical project as a new project: provide a name and path, and a new vertical project will be set up in an R studio project.
}
\examples{
\dontrun{
init_vertical_project()
}


}
