% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{update_yml}
\alias{update_yml}
\title{Update _pkgdown.yml for Rmds in vertical folders}
\usage{
update_yml(vertical_folder, docs_folder, yml_component, tab_title)
}
\arguments{
\item{vertical_folder}{character, name of a vertical folder, e.g., "vignettes"}

\item{docs_folder}{character, name of folder in docs, e.g., "articles"}

\item{yml_component}{character, name of the corresponding yml component, e.g., "articles"}

\item{tab_title}{character, the title of the component tab written to the website, e.g., "Supplementary"}
}
\value{
file, an updated _pkgdown.yml file
}
\description{
This function scrapes a folder and all sub-folders for R markdown documents, and then updates the a designated yml component in \verb{_pkgdown.yml} with a new list of items. The list of items in shown in the component's tab on the website, and the title of the tab is set to \code{tab_title}. In the tab, the titles of each .rmd in the main folder are added; then, for each sub-folder the sub-folder name is added (as a text seperator in the tab), followed by the titles for each .rmd in the sub-folder. This function is run automatically by \code{build_vertical()}.
}
