% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_project.R
\name{vertical_project}
\alias{vertical_project}
\title{Initialize a **vertical** project as RStudio project template}
\usage{
vertical_project(path, ...)
}
\arguments{
\item{path}{Where to create the project. This must be a valid R package name.}

\item{...}{Not used.}
}
\description{
This function is called when a **vertical** R project is created.
}
