% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{vertical_project}
\alias{vertical_project}
\title{Initialize \strong{vertical} project from RStudio project template}
\usage{
vertical_project(path, ...)
}
\arguments{
\item{path}{Where to create the project. This must be a valid R package name.}

\item{...}{Not used.}
}
\description{
\strong{NOT for use in the console}
}
\details{
\code{vertical_project()} is triggered by loading a new vertical project template from RStudio, and is a wrapper function to create a vertical project structure in the new project folder. The inputs  to this function are selections from the new project template window. Checkbox option for the template are defined in \code{inst/rstudio/templates/project/vertical_project.dcf}.
}
