% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{vertical_project}
\alias{vertical_project}
\title{Initialize a \strong{vertical} project}
\usage{
vertical_project(
  path,
  init_git = TRUE,
  init_data = TRUE,
  init_ms = TRUE,
  init_som = TRUE,
  init_slides = TRUE,
  init_poster = TRUE,
  init_exp = FALSE,
  ...
)
}
\arguments{
\item{path}{Name (and location) of package. Must be a valid R package name.}

\item{...}{Not used.}
}
\description{
This function is called when a \strong{vertical} R project is created.
}
